archive/v1-microservices/rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
structlog==23.2.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
qdrant-client==1.7.0
docling==1.0.0
docling-core==1.0.0
sentence-transformers==2.2.2
python-magic==0.4.27
pypdf==3.17.1
python-docx==1.1.0
python-pptx==0.6.23
unstructured==0.11.6
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1